refactor(MapView): Change results displayed on Marker
